Rename pkg/input/appcatalog to pkg/input/helmrepoindex
