+++ pkg/input/helmrepoindex/testdata/giantswarm.yaml
